Just bit refactor.
